Add typing to entity filters
